no msg

+++ stm32/src/include/hygrometer.h
@@ -2,6 +2,6 @@
 #define __HYGROMETER_H__
 
 void hygrometer_init();
-void hygrometer_read(int* temp, int* humidity);
+void hygrometer_read(short* temp, short* humidity);
 
 #endif
